requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.12
django-appconf==1.0.2
django-bootstrap3==9.1.0
django-ckeditor==5.4.0
django-decouple==2.1
django-js-asset==1.1.0
django-taggit==0.22.2
django-uuslug==1.1.8
django-versatileimagefield==1.9
gunicorn==19.7.1
idna==2.6
jsmin==2.2.2
olefile==0.44
pilkit==2.0
Pillow==5.0.0
python-decouple==3.1
python-slugify==1.2.5
pytz==2018.4
requests==2.18.4
six==1.11.0
static3==0.7.0
Unidecode==1.0.22
urllib3==1.22
